separated volume tagging test to a separate file
