+++ test/test_parallel_consistency_static.py
@@ -97,23 +97,3 @@ def test_parallel_consistency(
         compute_l2_error_with_reference_data,
         tolerances,
     )
-
-
-def test_parallel_consistency_volume_tagging(
-    request,
-    temp_results_path,
-    load_vtu_file,
-    get_cell_centroid_point_dofs,
-    compute_l2_error_with_reference_data,
-    tolerances,
-):
-    """Test parallel consistency for volume-tagged output."""
-    _run_parallel_consistency(
-        request.config.getoption("domain_dimension"),
-        "parallel_volume_output",
-        temp_results_path,
-        load_vtu_file,
-        get_cell_centroid_point_dofs,
-        compute_l2_error_with_reference_data,
-        tolerances,
-    )
